refactor to use react memo
